+++ .gitignore
@@ -23,3 +23,4 @@ bin/
 dump/
 obj/
 results/*
+py_scripts/
