+++ Makefile
@@ -10,9 +10,24 @@ SRC_DIR = src
 OBJ_DIR = obj
 BIN_DIR = bin
 
-# Targets
+# Source files
+SRCS = $(SRC_DIR)/main.cpp \
+       $(SRC_DIR)/nms.cpp \
+       $(SRC_DIR)/preprocess.cpp \
+       $(SRC_DIR)/inference.cpp \
+       $(SRC_DIR)/postprocess.cpp \
+       $(SRC_DIR)/util.cpp
+
+# Object files
+OBJS = $(OBJ_DIR)/main.o \
+       $(OBJ_DIR)/nms.o \
+       $(OBJ_DIR)/preprocess.o \
+       $(OBJ_DIR)/inference.o \
+       $(OBJ_DIR)/postprocess.o \
+       $(OBJ_DIR)/util.o
+
+# Target executable
 TARGET = $(BIN_DIR)/cudacatify
-OBJS = $(OBJ_DIR)/main.o
 
 # Default target
 all: $(TARGET)
@@ -28,10 +43,10 @@ $(BIN_DIR):
 $(TARGET): $(OBJS) | $(BIN_DIR)
 	$(CXX) $(OBJS) -o $(TARGET) $(LDFLAGS)
 
-# Compile C++ source file (main.cpp is in src directory)
-$(OBJ_DIR)/main.o: $(SRC_DIR)/main.cpp | $(OBJ_DIR)
-	$(CXX) -c $(SRC_DIR)/main.cpp -o $(OBJ_DIR)/main.o $(CXXFLAGS)
+# Rule to compile each source file into an object file
+$(OBJ_DIR)/%.o: $(SRC_DIR)/%.cpp | $(OBJ_DIR)
+	$(CXX) -c $< -o $@ $(CXXFLAGS)
 
 # Clean build artifacts
 clean:
-	rm -f $(OBJ_DIR)/*.o $(BIN_DIR)/cudacatify
+	rm -f $(OBJ_DIR)/*.o $(TARGET)
